backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pydantic
python-dotenv
pypdf
python-docx
openpyxl
playwright